Refactored Backend APIs
