+++ backend/src/api-docs/openAPIDocumentGenerator.ts
@@ -1,10 +1,11 @@
 import { OpenApiGeneratorV3, OpenAPIRegistry } from '@asteasolutions/zod-to-openapi'
 
 import { healthCheckRegistry } from '@/api/healthCheck/healthCheck.routes'
-import { templateRegistry } from '@/api/templates/template.controller'
+import { templateRegistry } from '@/api/templates/template.routes'
+import { sideBarRegistry } from '@/api/sideBar/sideBar.routes'
 
 export function generateOpenAPIDocument() {
-    const registry = new OpenAPIRegistry([healthCheckRegistry, templateRegistry])
+    const registry = new OpenAPIRegistry([healthCheckRegistry, templateRegistry, sideBarRegistry])
     const generator = new OpenApiGeneratorV3(registry.definitions)
 
     return generator.generateDocument({
