+++ backend/src/server.ts
@@ -7,7 +7,8 @@ import helmet from 'helmet'
 
 // Controllers (route handlers)
 import { healthCheckRouter } from './api/healthCheck/healthCheck.routes'
-import { templateController } from './api/templates/template.controller'
+import { sideBarRouter } from './api/sideBar/sideBar.routes'
+import { templateController } from './api/templates/template.routes'
 import { openAPIRouter } from './api-docs/openAPIRouter'
 // Middleware
 // Helpers
@@ -25,6 +26,7 @@ app.set('port', PORT!)
 
 // Primary app routes
 app.use('/v1/health-check', healthCheckRouter)
+app.use('/v1/sidebar', sideBarRouter)
 app.use('/v1/template', templateController)
 
 // Swagger UI
